Add min_point parameter for obstacle_layer
this is to add a parameter for marking obstacles on costmap, default
behavior is for each point marking obstacle on costmap ceil, this
parameter allows user to set a min value for marking obstacle on that
ceil. used when depth sensor reports random point sometimes.

default value is 1, as original code, max is 255.